routes file added & working
